move RDBTracingListenerConfiguration
