+++ elastic-job-lite-tracing/elastic-job-lite-tracing-rdb/src/main/resources/META-INF/services/org.apache.shardingsphere.elasticjob.lite.tracing.listener.TracingListenerConfiguration
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.lite.tracing.rdb.config.RDBTracingListenerConfiguration
+org.apache.shardingsphere.elasticjob.lite.tracing.rdb.listener.RDBTracingListenerConfiguration
